Treat regexp literals as strings in proper contexts
Mark regexp literals as strings by modifying the syntax table
properties, to make them play more nicely with sexp navigation.
Notably, forward-sexp will not move across the entire regexp literal.

Also limit the fontification and treatment of regexp literals to
contexts in which they are allowed, that is, in cases or selectors, node
names, and after match operators.