Add support for indentation of argument lists
Use the same indentation code as arrays. Move the array indentation
into a function and modify it to also handle argument lists, including
class parameter lists. Closing braces of arrays is now also aligned
correctly when on a line by itself.

Example:

class foo ($bar,
           $bar
) {
  $foo = foo($bar,
             $bar
	    )
  $foo = [$bar,
          $bar,
         ]
}

This does not change the indentation of argument lists with a newline
after the opening parenthesis:

class foo (
  $bar,
) {
  $foo = foo(
    $bar
  )
}