Ignore lines ending with a comment
Ignore lines ending with a comment when searching for indentation
hints. This is to prevent special characters in comments from changing
the indentation level.

One situation when this happens is when commenting out chunks of
Puppet code. The following:

  # class foo {
  #   foo { 'bar':
  #     foo => bar;
  #   }
  # }

was indented like this:

  # class foo {
    #   foo { 'bar':
      #     foo => bar;
    #   }
  # }

Before this commit a comment at the end of a line already prevented
indentation from working correctly in most cases. This is due to the
various regular expressions only allowing whitespace after the special
characters they search for.

One exception is that comments after an opening curly brace used to
work. With this commit the following

  foo { # comment about foo
    $foo = $bar
  }

now incorrectly is indented like this

  foo { # comment about foo
  $foo = $bar
  }