Fix indentation following array with nested hash (#125)
* Fix the indentation of a line which follows an array that contains a
hash as one of its elements.

https://github.com/voxpupuli/puppet-mode/issues/108#issuecomment-399312397

puppet-indent-line currently scans backwards in the buffer, line by
line, and if it finds a closing (brace or paren) (optionally followed
by a comma) then it records the current indentation level and returns
that as cur-indent.

If we nest data structures, this heuristic can break in the event that
the array list contains braces to denote a hash.

I propose to fix this issue by adding a closing bracket to the list of
punctuation recognised by the heuristic - the backwards search of the
buffer then terminates at the "]," which denotes the end of the array,
and never gets confused by closing braces or closing parens within the
array block.

* Fix regexp for detecting brace, paren or bracket.

Correction to the regexp modified in commit d8386be; that attempt
basically stopped the regexp from matching at all, which happened to
have the desired effect on my particular test case, but broke other
indentation. This RE actually matches the intent declared in the
comment.

* Handle closing bracket similar to closing paren.

Fixes a failure (introduced by commit e42b90a) of Travis-CI test
"puppet-indent-line/array".