Add skeletons for types & statements (closes #31)
These skeletons insert common types and statements using mnemonic
keybinding.  The "C-c C-t" prefix is used for types and the "C-c C-k"
prefix is used for keywords.  Some of the statement skeletons ("if",
"elsif", "else" & "unless") wrap around the region if it is defined.
Turning an existing block of code into a conditional block of code is
now an easy task with braces and indentation handled automatically.